manual net cleaned - 'while' and 'break' instead of 'sys.exit'
