+++ conf_wizard_net.py
@@ -1,6 +1,6 @@
 import os
 from modules import load_config, clear_the_screen, logo_top, Bcolors
-from net_and_ap import net_and_ap_conf
+from net_and_ap_man_conf import net_and_ap_conf
 
 
 def confirm_auto_hotspot(config):
